requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
celery==4.4.2
Django==3.0.5
django-filter==2.2.0
django-redis==4.11.0
django-taggit==1.2.0
djangorestframework==3.11.0
gunicorn==20.0.4
kombu==4.6.8
Markdown==3.2.1
mysqlclient==1.4.6
PyJWT==1.7.1
pytz==2020.1
redis==3.5.0
sqlparse==0.3.1
vine==1.3.0
